Fix broken document header in norms
Changes in the behavior of the header in the caselaw space of the app
were incompatible with the norms page. This commit fixes the norms
header for now. Unfortunately this is very hacky (sorry). We should
refactor the entire layout and header situation soon as the different
parts of the app are starting to diverge and it's becoming harder to
maintain all of this as shared components.

RISDEV-0000

From commit: https://github.com/digitalservicebund/ris-backend-service/commit/71d651fc6efb6d1c9850465b2b2bec83fa0a73fc