Test(norms): improve chips interactions for e2e
As already mentioned for the unit tests, the interactions with the chips
input isn't too easy. Therefore we introduced a temporary solution that
uses test identifier properties to help addressing everything properly.
This has now been applied to the end-to-end tests too.

In the past we have been really lucky that this actually worked. We now
realized that the verification of chips was basically randomly
searching for any visible chip with this text over the whole page
anywhere.
Now we verify very accurately that we check the chips for the specific
input we are interested in. We also verify the exact count and order of
them too.
This also uncovered one issue where our expected values have been wrong,
though the tests have passed in the passed due to the above described
(extremely) relaxed checks.

RISDEV-1969

From commit: https://github.com/digitalservicebund/ris-backend-service/commit/500495947c78c007a9ba37e94a49a1e3c784e3dc